+++ CHANGELOG.md
@@ -5,6 +5,7 @@ untagged
 --------
 
 * Add method for get count messages from queue.
+* Add delay system (allow publish messages with delay).
 
 v1.0.0
 ------